Break profileDetails intro view and edit components
