+++ app/components/EditProfile/profileDetails.css
@@ -1,0 +1,280 @@
+.component.editProfile {
+  width: 100%;
+  height: auto;
+  background-color: var(--color-primary-light);
+  border-radius: var(--border-radius-l);
+  margin-bottom: var(--padding-l);
+  box-shadow: var(--primary-shadow);
+  box-sizing: border-box;
+
+  & .editWrapper {
+    width: 100%;
+    box-sizing: border-box;
+    padding: var(--padding-l);
+    margin-bottom: var(--padding-l);
+
+    & fieldset {
+      padding-bottom: var(--padding-m);
+    }
+
+    & .fileInput {
+      margin-bottom: var(--padding-s);
+    }
+
+    & .descriptionInput {
+      height: var(--right-sidebar);
+      resize: vertical;
+    }
+
+    & .socialAccountInput {
+      display: flex;
+      flex-flow: row nowrap;
+      padding: 0;
+
+      & .input {
+        flex-grow: 1;
+        flex-shrink: 1;
+        margin-right: var(--padding-s);
+      }
+
+      & .removeButton {
+        flex-grow: 0;
+        flex-shrink: 1;
+        display: flex;
+        flex-flow: row;
+        justify-content: center;
+        align-items: center;
+        font-size: 1.1rem;
+        color: gray;
+        border: color(srgb 0.935 0.8859 0.8696) solid 1px;
+        background-color: color(srgb 0.985 0.9586 0.9505);
+        padding: 0 11px;
+
+        &:hover {
+          cursor: pointer;
+          color: var(--color-secondary);
+          background-color: var(--white-see-through);
+        }
+      }
+    }
+  
+    & .button {
+      margin-bottom: var(--padding-s);
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      flex-direction: column;
+      padding: var(--size-h3);
+    }
+
+    & .showFee {
+      font-size: var(--padding-xs);
+      line-height: 14px;
+    }
+  }
+}
+
+.component.viewProfile {
+  width: 100%;
+  height: auto;
+  background-color: var(--color-primary-light);
+  border-radius: var(--border-radius-l);
+  margin-bottom: var(--padding-l);
+  box-shadow: var(--primary-shadow);
+  box-sizing: border-box;
+
+  & .viewWrapper {
+    display: flex;
+    flex-flow: column nowrap;
+    align-items: center;
+    flex-direction: column;
+    position: relative;
+
+    & .bubblesPattern {
+      position: absolute;
+      bottom: 0;
+      right: 3rem;
+      width: 292px;
+      height: 43px;
+      z-index: 0;
+
+      & img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+      }
+    }
+
+    & .profileAvatar {
+      position: relative;
+      width: 10rem;
+      height: 10rem;
+      overflow: hidden;
+      border-radius: var(--border-radius-x);
+      margin-bottom: var(--padding-l);
+      margin-top: var(--padding-x);
+
+      & img {
+        height: 100%;
+        object-fit: cover;
+        width: 100%;
+      }
+    }
+
+    & .profileName {
+      width: 100%;
+      text-align: center;
+      padding: var(--padding-xs);
+
+      & h2 {
+        display: inline-block;
+        max-width: 100%;
+        margin: 0;
+        color: var(--color-secondary);
+        font-weight: var(--font-weight-black);
+        word-break: break-all;
+      }
+    }
+
+    & .bioContainer {
+      padding-top: var(--padding-m);
+      padding-inline: var(--padding-m);
+
+      & h4,
+      & p {
+        color: var(--gray);
+      }
+    }
+
+    & .socialMediaAccounts {
+      align-self: baseline;
+      padding: var(--padding-l) var(--padding-m);
+      margin: 0 auto var(--padding-l);
+
+      & ul {
+        display: flex;
+        gap: 2rem;
+        margin-bottom: var(--padding-m);
+
+        & li {
+          flex-grow: 1;
+          flex-shrink: 0;
+
+          & a {
+            display: flex;
+            flex-direction: column;
+            
+            &[href=''] {
+              pointer-events: none;
+            }
+
+            & span {
+              color: var(--gray);
+              text-align: center;
+            }
+          }
+
+          & .iconBox {
+            position: relative;
+            background-color: var(--color-primary);
+            text-align: center;
+            width: 67px;
+            height: 65px;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            border-radius: var(--border-radius-m);
+            margin-bottom: var(--padding-s);
+
+            & .icon.twitter {
+              color: #00acee;
+            }
+
+            & .icon.youtube {
+              color: red;
+            }
+
+            & .icon.instagram {
+              background-image: radial-gradient(
+                circle at 30% 107%,
+                #fdf497 0%,
+                #fdf497 5%,
+                #fd5949 45%,
+                #d6249f 60%,
+                #285aeb 90%
+              );
+              background-clip: text;
+              color: transparent;
+            }
+
+            & i {
+              font-size: 33px;
+              min-width: 100%;
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+@media (min-width: 480px) {
+  .component.viewProfile {
+    & .viewWrapper {
+      flex-flow: row wrap;
+
+      & .profileAvatar {
+        margin-left: var(--padding-m);
+      }
+
+      & .profileName {
+        width: 45%;
+      }
+
+      & .bioContainer {
+        width: 100%;
+      }
+
+      & .social {
+        align-self: center;
+      }
+    }
+  }
+}
+
+@media (min-width: 900px) {
+  .component.viewProfile {
+    & .viewWrapper {
+      & .profileName {
+        width: 39%;
+      }
+    }
+  }
+}
+
+@media (min-width: 1280px) {
+  .component.viewProfile {
+    & .viewWrapper {
+      justify-content: space-around;
+
+      & .profileAvatar {
+        width: 15.5rem;
+        height: 15.5rem;
+      }
+
+      & .profileName {
+        align-self: self-start;
+        padding-top: var(--box-size-xl);
+        text-align: left;
+      }
+
+      & .bioContainer {
+        width: 44%;
+        order: 4;
+        align-self: flex-start;
+        padding-top: 0;
+        margin-top: calc(var(--box-size-xxl) * -1);
+      }
+    }
+  }
+}
